Some cleanup and refactoring
