+++ src/lib/crypto/ec.cpp
@@ -32,9 +32,9 @@
 #include "utils.h"
 #include "mem.h"
 #include "bn.h"
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
-#include "x25519.h"
-#include "ed25519.h"
+#if defined(ENABLE_CRYPTO_REFRESH)
+#include "x25519_x448.h"
+#include "ed25519_ed448.h"
 #include "botan/bigint.h"
 #include "botan/ecdh.h"
 #include <cassert>
@@ -193,7 +193,7 @@ end:
     return ret;
 }
 
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
 static bool
 is_generic_prime_curve(pgp_curve_t curve)
 {
