+++ src/lib/crypto/exdsa_ecdhkem.cpp
@@ -26,7 +26,7 @@
 
 #include "exdsa_ecdhkem.h"
 #include "ecdh.h"
-#include "ed25519.h"
+#include "ed25519_ed448.h"
 #include "ecdsa.h"
 #include "ec.h"
 #include "types.h"
