+++ src/lib/crypto/rng.cpp
@@ -35,7 +35,7 @@ RNG::RNG(Type type)
     if (botan_rng_init(&botan_rng, type == Type::DRBG ? "user" : NULL)) {
         throw rnp::rnp_exception(RNP_ERROR_RNG);
     }
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
     if (type == Type::DRBG) {
         botan_rng_obj.reset(new Botan::AutoSeeded_RNG);
     } else {
@@ -64,7 +64,7 @@ RNG::handle()
     return botan_rng;
 }
 
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
 Botan::RandomNumberGenerator *
 RNG::obj() const
 {
