+++ src/librepgp/stream-parse.cpp
@@ -1567,7 +1567,7 @@ encrypted_start_aead(pgp_source_encrypted_param_t *param, pgp_symm_alg_t alg, ui
 #endif
 }
 
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
 /* The crypto refresh mandates that for a X25519/X448 PKESKv3, AES MUST be used.
    The same is true for the PQC algorithms
  */
@@ -1587,9 +1587,7 @@ do_enforce_aes_v3pkesk(pgp_pubkey_alg_t alg)
     case PGP_PKA_KYBER1024_BP384:
         FALLTHROUGH_STATEMENT;
 #endif
-#if defined(ENABLE_CRYPTO_REFRESH)
     case PGP_PKA_X25519:
-#endif
         return true;
     default:
         return false;
@@ -1632,7 +1630,7 @@ encrypted_try_key(pgp_source_encrypted_param_t *param,
     }
 #endif
 
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
     /* check that AES is used when mandated by the standard */
     if (do_enforce_aes_v3pkesk(sesskey.alg) && sesskey.version == PGP_PKSK_V3) {
         switch (sesskey.salg) {
@@ -1664,7 +1662,7 @@ encrypted_try_key(pgp_source_encrypted_param_t *param,
     uint8_t *decbuf_sesskey = decbuf.data();
     size_t   decbuf_sesskey_len = declen;
     size_t   keylen;
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
     if (do_encrypt_pkesk_v3_alg_id(sesskey.alg))
 #endif
     {
@@ -1680,7 +1678,7 @@ encrypted_try_key(pgp_source_encrypted_param_t *param,
             return false;
         }
 
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
         size_t alg_id_bytes = do_encrypt_pkesk_v3_alg_id(sesskey.alg) ? 1 : 0;
         size_t checksum_bytes = have_pkesk_checksum(sesskey.alg) ? 2 : 0;
 #else
@@ -1714,7 +1712,7 @@ encrypted_try_key(pgp_source_encrypted_param_t *param,
                   std::vector<uint8_t>(decbuf_sesskey, decbuf_sesskey + keylen));
 #endif
 
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
     if (have_pkesk_checksum(sesskey.alg))
 #endif
     {
@@ -1731,7 +1729,7 @@ encrypted_try_key(pgp_source_encrypted_param_t *param,
         }
     }
 
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
     if (sesskey.version == PGP_PKSK_V3)
 #endif
     {
