+++ src/tests/exdsa_ecdhkem.cpp
@@ -24,7 +24,7 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#if defined(ENABLE_CRYPTO_REFRESH) || defined(ENABLE_PQC)
+#if defined(ENABLE_CRYPTO_REFRESH)
 
 #include "rnp_tests.h"
 #include "crypto/exdsa_ecdhkem.h"
